+++ frontend/src/lib/components/accounts/RenameSubAccountAction.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
   import { i18n } from "$lib/stores/i18n";
   import { createEventDispatcher, getContext } from "svelte";
-  import { renameSubAccount } from "$lib/services/accounts.services";
+  import { renameSubAccount } from "$lib/services/icp-accounts.services";
   import { startBusy, stopBusy } from "$lib/stores/busy.store";
   import { busy } from "@dfinity/gix-components";
   import type { Account } from "$lib/types/account";
