+++ frontend/src/lib/modals/sns/sale/ParticipateSwapModal.svelte
@@ -42,7 +42,7 @@
   import {
     cancelPollAccounts,
     pollAccounts,
-  } from "$lib/services/accounts.services";
+  } from "$lib/services/icp-accounts.services";
   import type { TransactionInit } from "$lib/types/transaction";
 
   onMount(() => {
