+++ frontend/src/lib/pages/NnsAccounts.svelte
@@ -10,7 +10,7 @@
   import {
     cancelPollAccounts,
     pollAccounts,
-  } from "$lib/services/accounts.services";
+  } from "$lib/services/icp-accounts.services";
   import { ICPToken } from "@dfinity/utils";
 
   export let goToWallet: (account: Account) => Promise<void>;
