+++ frontend/src/lib/services/dev.services.ts
@@ -12,7 +12,7 @@ import {
 } from "$lib/stores/sns-accounts.store";
 import type { Principal } from "@dfinity/principal";
 import { get } from "svelte/store";
-import { syncAccounts } from "./accounts.services";
+import { syncAccounts } from "./icp-accounts.services";
 import { loadSnsAccounts } from "./sns-accounts.services";
 
 export const getTestBalance = getTestAccountBalance;
